Increase precision of time slices; give option to use slower slices
Some reports of shuddering/chattering type motion appear to be caused
by computers that are either slow or busy (and acting slow). We are
adding a GUI option to use wider time slices (50 ms), that can be
activated in the event that shuddering motion is observed. Also
decreasing default (not-slowed-down) time slices to 25 ms wide (from 30
ms), giving a 20% improvement in smoothness.