Catch rare issue with excessive speed
Heuristic fix for case of triangle segment profile, where Vmax is
greater than speedLimit. Can occur near maximum speeds, when a pen-up
move is of exactly the right length that it's long enough for
acceleration to full speed, but not long enough to also have at least
25 ms of "cruising" time at full speed.

Fix is by reducing acceleration slightly to keep the peak speed below
our maximum.  It may be preferable to rephrase this as a trapezoid
profile with a non-maximal speed.