Rand seed, temporary changes to SVG, preview lines
* Add a random seed to the set of "WCB" data stored when a plot is
paused. This will allow a plot that includes some random components to
be paused and resumed later.
* Revert the SVG to the original document at the end of a plot, before
saving any plot-in-progress data. This will allow us to muck with the
SVG data itself (font substitution or rendering, alteration of object
order, adding/removing objects and so forth) during the plotting
process, without needing to also work hard to preserve the original
document.
* Put pen-down preview layer above pen-up preview layer, for better
clarity.
* Improve quality of preview lines with rounded joining and endcaps.