Address bug with short
Sufficiently short pen-up moves (smaller than a typical pen width) were previously skipped, in the interest of time savings.
This leads to a small but real inaccuracy in the starting point of the next line. We now replace those short pen-up moves with a pen-down move.
Also add direct variable to indicate when the script has been called externally.