version 3.9.3
Add timeout on webhooks
Improved error handling on webhooks
Bug fix for printing a specific layer from a plob that contains multiple layers, which can occur when generating the plob in the API and then using layers mode.
Alter value for "standard" pen-lift rate. (Not the default value; a slower value called "standard".)